requirements.txt

langchain
langgraph
langchain-core
langchain-community
langchain-groq
arxiv
wikipedia
tavily-python  
python-dotenv
pydantic
typing_extensions
ipykernel
IPython
graphviz  
requests  
streamlit
